martian: move copySync to separate file
