+++ internal/martian/proxy.go
@@ -430,44 +430,6 @@ func shouldTerminateTLS(req *http.Request) bool {
 	return b
 }
 
-func drainBuffer(w io.Writer, r *bufio.Reader) error {
-	if n := r.Buffered(); n > 0 {
-		rbuf, err := r.Peek(n)
-		if err != nil {
-			return err
-		}
-		w.Write(rbuf)
-	}
-	return nil
-}
-
-var copyBufPool = sync.Pool{
-	New: func() any {
-		b := make([]byte, 32*1024)
-		return &b
-	},
-}
-
-func copySync(ctx context.Context, name string, w io.Writer, r io.Reader, donec chan<- bool) {
-	bufp := copyBufPool.Get().(*[]byte) //nolint:forcetypeassert // It's *[]byte.
-	buf := *bufp
-	defer copyBufPool.Put(bufp)
-
-	if _, err := io.CopyBuffer(w, r, buf); err != nil && !isClosedConnError(err) {
-		log.Errorf(ctx, "failed to copy %s tunnel: %v", name, err)
-	}
-	if cw, ok := asCloseWriter(w); ok {
-		cw.CloseWrite()
-	} else if pw, ok := w.(*io.PipeWriter); ok {
-		pw.Close()
-	} else {
-		log.Errorf(ctx, "cannot close write side of %s tunnel (%T)", name, w)
-	}
-
-	log.Debugf(ctx, "%s tunnel finished copying", name)
-	donec <- true
-}
-
 // A peekedConn subverts the net.Conn.Read implementation, primarily so that
 // sniffed bytes can be transparently prepended.
 type peekedConn struct {
